Reapply changes pertaining to analyzer parametrization (#146)
* [ENH] Hashtag update and add tests (#121)

* [ENH] update interface to rely on post column, add users col

* [ENH] update gini() to operate on pl.Series

* [ENH] wrap analysis code into hashtag_analyzer()

* [ENH] check for time column dtype

* [ENH] fix column naming, set defaul window to 12h

* [ENH] add data and tests for hashtag_analyzer

* [ENH] add tests for gini()

* [MAINT] explicitly specify return_dtype

* [MAINT] cleanup

* fix: hide unsupported hashtag analysis export formats

* test tutorial comment cleanup

* fix: Capitalization typo

---------

Co-authored-by: DeanEby <ebyd21@gmail.com>

* feat: analyzer parameterization (#126)

This change introduces analyzer parameterization. Parameters can be defined
using `AnalyzerParam` and queried via the analyzer/web presenter context
objects. The CLI is updated to include a step where parameters are chosen,
and they are reflected back when the analysis is viewed. The test helpers
are also updated to include parameter provision.

The example analyzer, which performs simple character count, is updated
to include a single parameter, namely `fudge_factor` that offsets the
character count.

These parameter types are supported:
- `IntegerParam` (produces `int` value)
- `TimeBinningParam` (produces `TimeBinningValue`, essentially an object
  with attributes `unit` and `amount`.

The `TimeBinningParam` is currently unused, but it is defined in anticipation
that it will be used by the hashtag tests. As such, a convenient method that
produces a polars `truncate` expression is added to `TimeBinningValue` so
it can just be dropped into the analyzer later.

* [FEAT] Add time window parameter selection to hashtags test (#129)

* add time bining parameter to hashtag analyzer
* add default parameter for tester
* de-indent help text, remove reference to polars expr

---------

Co-authored-by: DeanEby <ebyd21@gmail.com>
Co-authored-by: soul-codes <40335030+soul-codes@users.noreply.github.com>